Committing an up-to-date version of my engine. It now compiles and a basic version of it runs in 64-bit mode. Once it is up to date fully, I will remove the 32-bit option entirely

+++ Source/private/Memory/Heap.cpp
@@ -43,7 +43,7 @@ Heap::Heap(const char* heapName, size_t heapSize, Heap* nextHeap, HANDLE heapHan
 	}
 
 	mInfo.TotalSize = heapSize;
-	mInfo.StartAddr = (unsigned int)this;
+	mInfo.StartAddr = (uint64)this;
 	mInfo.EndAddr = mInfo.StartAddr + heapSize;
 	
 	// Name initialization
